Extracted FileFromUrl before moving it to another package.
